Nałożona funkcja MC2D

+++ main.cpp
@@ -10,9 +10,16 @@ using namespace std;
 
 // aproksymacja całki gausowskiej od 0 do 1
 
+// zamienić dla ogólnej funkcji N wymiarowej
+
 double g(double x)
 {
     return sin(x);
+}
+
+double f(double x1, double x2)
+{
+    return exp(x1+x2);
 }
 
 tuple <double,double> MC1D(int N, double a, double b)
@@ -77,15 +84,46 @@ void app1sam(int N, double a, double b, double intVal)
     file.close();
 }
 
+tuple <double,double> MC2D(int N, double a[2], double b[2])
+{
+    double V = 0; 
+    double S = 0;
+    double xi1;
+    double xi2;
+    for(int i = 0; i < N; i++)
+    {
+        xi1 = rand()/(RAND_MAX + 1.);
+        xi2 = rand()/(RAND_MAX + 1.);
+        S += f(b[1]+(b[0]-b[1])*xi1,a[1]+(a[0]-a[1])*xi2);
+        V += pow(f(xi1,xi2),2);
+    }
+    double MC = S*((b[1]-b[0])*(a[1]-a[0])/(double)N);
+    return {MC,0}; 
+}
+
 int main()
 {
     srand(time(NULL));
 
-    approx1D(1e6,0,M_PI,2);
-    app1sam(1e4,0,M_PI,2);
+    //approx1D(1e6,0,M_PI,2);
+    //app1sam(1e4,0,M_PI,2);
     
-    system("gnuplot < script.gnu");
-    system("rm data.dat && rm data2.dat");
+    //system("gnuplot < script.gnu");
+
+    double a[] = {0,1};
+    double b[] = {0,1}; 
+
+    cout << get<0>(MC2D(1000000,a,b)) << endl;
+    
+
+    ifstream ifile;
+    ifile.open("data.dat");
+
+    if(ifile)
+    {
+        ifile.close();
+        system("rm data.dat && rm data2.dat");
+    }
 
     return 0;
 }
